# api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserBody"}}}},"responses":{"200":{"description":"User has been logged in.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserResponse"}}}}}}},"/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenBody"}}}},"responses":{"200":{"description":"Generates a new access token.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenResponse"}}}}}}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserBody"}}}},"responses":{"201":{"description":"User has been registered.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserResponse"}}}}}}},"/users":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/users/{username}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/users/profile":{"put":{"operationId":"UsersController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}}}},"/users/me":{"get":{"operationId":"UsersController_getProfile","parameters":[],"responses":{"200":{"description":"Returns the logged-in user.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}}},"info":{"title":"Domicile API","description":"Domicile API","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginUserBody":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["id","username","firstName","lastName"]},"LoginUserResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"},"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["user","accessToken","refreshToken"]},"RefreshTokenBody":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"RefreshTokenResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"},"accessToken":{"type":"string"}},"required":["user","accessToken"]},"RegisterUserBody":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"RegisterUserResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"},"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["user","accessToken","refreshToken"]},"UpdateUserDto":{"type":"object","properties":{"username":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["username","firstName","lastName"]}}}}